pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secure-aggregation-fl"
version = "0.1.0"
description = "Modular secure aggregation prototype with topology, crypto, and config layers."
authors = [{ name = "Secure Aggregation Team" }]
requires-python = ">=3.10"
dependencies = [
  "cryptography>=42.0.0,<44.0.0",
  "fastapi>=0.110.0",
  "uvicorn>=0.23.0",
  "httpx>=0.27.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
test = [
  "pytest>=7.4",
  "pyyaml>=6.0",
]
mnist = [
  "torch>=2.2.0",
  "torchvision>=0.17.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["src"]
